# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Movie Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Movie {
  tmdb_id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  /** @nullable */
  release_date?: string | null;
  /** @nullable */
  vote_average?: number | null;
  genre_ids?: number[];
}

export type MovieDetailsGenresItem = {
  id: number;
  name: string;
};

export type MovieDetailsCastItem = {
  id: number;
  name: string;
  character: string;
  /** @nullable */
  profile_path?: string | null;
};

export interface MovieDetails {
  tmdb_id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  /** @nullable */
  release_date?: string | null;
  /** @nullable */
  vote_average?: number | null;
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  tagline?: string | null;
  genres?: MovieDetailsGenresItem[];
  cast?: MovieDetailsCastItem[];
}

export interface MovieSearchResult {
  results: Movie[];
  total_results: number;
  total_pages: number;
  page?: number;
}

export type ListEntryStatus = typeof ListEntryStatus[keyof typeof ListEntryStatus];


export const ListEntryStatus = {
  watched: 'watched',
  want_to_see: 'want_to_see',
} as const;

export interface ListEntry {
  id: string;
  user_id: string;
  tmdb_id: number;
  title: string;
  /** @nullable */
  poster_path?: string | null;
  status: ListEntryStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  created_at: string;
  updated_at?: string;
  genre_ids?: number[];
}

export type ListEntryInputStatus = typeof ListEntryInputStatus[keyof typeof ListEntryInputStatus];


export const ListEntryInputStatus = {
  watched: 'watched',
  want_to_see: 'want_to_see',
} as const;

export interface ListEntryInput {
  user_id: string;
  tmdb_id: number;
  title: string;
  /** @nullable */
  poster_path?: string | null;
  status: ListEntryInputStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  genre_ids?: number[];
}

export type ListEntryUpdateStatus = typeof ListEntryUpdateStatus[keyof typeof ListEntryUpdateStatus];


export const ListEntryUpdateStatus = {
  watched: 'watched',
  want_to_see: 'want_to_see',
} as const;

export interface ListEntryUpdate {
  status?: ListEntryUpdateStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type UserStatsTopGenresItem = {
  genre_id: number;
  genre_name: string;
  count: number;
};

export interface UserStats {
  watched_count: number;
  want_to_see_count: number;
  /** @nullable */
  avg_rating: number | null;
  top_genres: UserStatsTopGenresItem[];
  total_movies?: number;
}

export type SearchMoviesParams = {
q: string;
page?: number;
};

export type GetTrendingMoviesParams = {
page?: number;
};

export type GetPopularMoviesParams = {
page?: number;
};

export type GetUserListParams = {
userId: string;
status?: GetUserListStatus;
};

export type GetUserListStatus = typeof GetUserListStatus[keyof typeof GetUserListStatus];


export const GetUserListStatus = {
  watched: 'watched',
  want_to_see: 'want_to_see',
} as const;

export type GetUserStatsParams = {
userId: string;
};

export type GetRecentActivityParams = {
userId: string;
limit?: number;
};