# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Movie Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: movies
    description: TMDB movie search and details
  - name: lists
    description: User movie lists (watched / want to see)
  - name: stats
    description: User statistics and activity

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /movies/search:
    get:
      operationId: searchMovies
      tags: [movies]
      summary: Search movies via TMDB
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieSearchResult"

  /movies/trending:
    get:
      operationId: getTrendingMovies
      tags: [movies]
      summary: Get trending movies
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Trending movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieSearchResult"

  /movies/popular:
    get:
      operationId: getPopularMovies
      tags: [movies]
      summary: Get popular movies
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Popular movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieSearchResult"

  /movies/{tmdbId}:
    get:
      operationId: getMovieDetails
      tags: [movies]
      summary: Get movie details from TMDB
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetails"

  /lists:
    get:
      operationId: getUserList
      tags: [lists]
      summary: Get all movies in the user's list
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [watched, want_to_see]
      responses:
        "200":
          description: User list entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListEntry"
    post:
      operationId: addToList
      tags: [lists]
      summary: Add a movie to the user's list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListEntryInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEntry"

  /lists/{id}:
    patch:
      operationId: updateListEntry
      tags: [lists]
      summary: Update a list entry (change status, rating, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListEntryUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEntry"
    delete:
      operationId: removeFromList
      tags: [lists]
      summary: Remove a movie from the user's list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /stats:
    get:
      operationId: getUserStats
      tags: [stats]
      summary: Get user statistics (counts, top genres, avg rating)
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

  /activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get user's recently added/watched movies
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListEntry"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Movie:
      type: object
      required: [tmdb_id, title]
      properties:
        tmdb_id:
          type: integer
        title:
          type: string
        overview:
          type: ["string", "null"]
        poster_path:
          type: ["string", "null"]
        backdrop_path:
          type: ["string", "null"]
        release_date:
          type: ["string", "null"]
        vote_average:
          type: ["number", "null"]
        genre_ids:
          type: array
          items:
            type: integer

    MovieDetails:
      type: object
      required: [tmdb_id, title]
      properties:
        tmdb_id:
          type: integer
        title:
          type: string
        overview:
          type: ["string", "null"]
        poster_path:
          type: ["string", "null"]
        backdrop_path:
          type: ["string", "null"]
        release_date:
          type: ["string", "null"]
        vote_average:
          type: ["number", "null"]
        runtime:
          type: ["integer", "null"]
        tagline:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
            required: [id, name]
        cast:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              character:
                type: string
              profile_path:
                type: ["string", "null"]
            required: [id, name, character]

    MovieSearchResult:
      type: object
      required: [results, total_results, total_pages]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Movie"
        total_results:
          type: integer
        total_pages:
          type: integer
        page:
          type: integer

    ListEntry:
      type: object
      required: [id, user_id, tmdb_id, title, status, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        tmdb_id:
          type: integer
        title:
          type: string
        poster_path:
          type: ["string", "null"]
        status:
          type: string
          enum: [watched, want_to_see]
        rating:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
        genre_ids:
          type: array
          items:
            type: integer

    ListEntryInput:
      type: object
      required: [user_id, tmdb_id, title, status]
      properties:
        user_id:
          type: string
        tmdb_id:
          type: integer
        title:
          type: string
        poster_path:
          type: ["string", "null"]
        status:
          type: string
          enum: [watched, want_to_see]
        rating:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        genre_ids:
          type: array
          items:
            type: integer

    ListEntryUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [watched, want_to_see]
        rating:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    UserStats:
      type: object
      required: [watched_count, want_to_see_count, avg_rating, top_genres]
      properties:
        watched_count:
          type: integer
        want_to_see_count:
          type: integer
        avg_rating:
          type: ["number", "null"]
        top_genres:
          type: array
          items:
            type: object
            properties:
              genre_id:
                type: integer
              genre_name:
                type: string
              count:
                type: integer
            required: [genre_id, genre_name, count]
        total_movies:
          type: integer