# lib/api-zod/src/generated/types/listEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Movie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ListEntryStatus } from './listEntryStatus';

export interface ListEntry {
  id: string;
  user_id: string;
  tmdb_id: number;
  title: string;
  /** @nullable */
  poster_path?: string | null;
  status: ListEntryStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  created_at: string;
  updated_at?: string;
  genre_ids?: number[];
}