# lib/api-zod/src/generated/types/listEntryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Movie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ListEntryInputStatus } from './listEntryInputStatus';

export interface ListEntryInput {
  user_id: string;
  tmdb_id: number;
  title: string;
  /** @nullable */
  poster_path?: string | null;
  status: ListEntryInputStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  genre_ids?: number[];
}