# lib/api-zod/src/generated/types/listEntryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Movie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ListEntryUpdateStatus } from './listEntryUpdateStatus';

export interface ListEntryUpdate {
  status?: ListEntryUpdateStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
}