# lib/api-zod/src/generated/types/movie.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Movie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Movie {
  tmdb_id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  /** @nullable */
  release_date?: string | null;
  /** @nullable */
  vote_average?: number | null;
  genre_ids?: number[];
}