# lib/api-zod/src/generated/types/movieDetails.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Movie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { MovieDetailsCastItem } from './movieDetailsCastItem';
import type { MovieDetailsGenresItem } from './movieDetailsGenresItem';

export interface MovieDetails {
  tmdb_id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  /** @nullable */
  release_date?: string | null;
  /** @nullable */
  vote_average?: number | null;
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  tagline?: string | null;
  genres?: MovieDetailsGenresItem[];
  cast?: MovieDetailsCastItem[];
}