# lib/api-zod/src/generated/types/movieSearchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Movie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Movie } from './movie';

export interface MovieSearchResult {
  results: Movie[];
  total_results: number;
  total_pages: number;
  page?: number;
}