# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Movie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatsTopGenresItem } from './userStatsTopGenresItem';

export interface UserStats {
  watched_count: number;
  want_to_see_count: number;
  /** @nullable */
  avg_rating: number | null;
  top_genres: UserStatsTopGenresItem[];
  total_movies?: number;
}